Rework the forms example to remove gender binary option
There are some arguments to be made that requiring gender on web forms
is usually unnecessary, and also that asking for it as a simple binary
can be exclusionary. I know this is just an example form, taken from
Wikipedia, but it irked me a bit so I rewrote it as a pizza shop
order.